feat: 테이블에 맞게 엔티티의 id 데이터 타입을 Integer 로 변경 및 리포지토리에 반영

+++ src/main/java/com/springboot/relationship/repository/HospitalRepository.java
@@ -3,6 +3,6 @@ package com.springboot.relationship.repository;
 import com.springboot.relationship.domain.entity.Hospital;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HospitalRepository extends JpaRepository<Hospital, Long> {
+public interface HospitalRepository extends JpaRepository<Hospital, Integer> {
 
 }
